refactor(moving files)[OK-95] Move undo redo command classes into a subdirectory
