+++ client/src/classes/undo-redo-commands/RemoveCommand.ts
