+++ client/src/errors/FabricObjectIdError.ts
